backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-dotenv==0.0.13
python-decouple==3.8
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
supabase==2.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
httpx==0.25.1
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0

